// printer-cloud-new/services/printer-air/types/document.ts
export interface IndexDocumentsAPIResponse {
  documents: Array<IndexDocument>;
  meta: APIMetaProperties;
}

export interface IndexDocument {
  id: number;
  originalFilename: string;
  status: DocumentStatus;
  description: string;
  location: string;
  directoryId: number;
  path: string;
  prn: string;
  previousParentPrn: string;
  createdAt: string;
  updatedAt: string;
  url: string;
  shared: boolean;
  shareableLink: boolean;
  updatedBy: UpdatedByIndexDocument | null;
}

export interface UpdatedByIndexDocument {
  id: number;
  name: string;
}
// ...
export type DocumentStatus = 'failed' | 'created' | 'enqueued' | 'processed';